+++ src/components/Navbar/Navbar.module.scss
@@ -1,5 +1,4 @@
 @use "src/styles/constants";
-@use "src/styles/utility";
 @use "src/styles/breakpoints";
 
 .emptySpacePlaceholder {
@@ -21,42 +20,5 @@
   transform: translateY(0 - constants.$navbar-height);
   @include breakpoints.tablet {
     transform: none;
-  }
-}
-
-.itemsContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  height: 100%;
-}
-
-.centerVertically {
-  @include utility.center-vertically;
-}
-
-.leftCTA {
-  display: flex;
-  margin-left: var(--spacing-medium);
-
-  @include breakpoints.tablet {
-    margin-left: var(--spacing-mega);
-  }
-}
-
-.rightCTA {
-  display: flex;
-  margin-right: var(--spacing-medium);
-
-  > button {
-    margin-left: var(--spacing-micro);
-  }
-
-  @include breakpoints.tablet {
-    margin-right: var(--spacing-mega);
-
-    > button {
-      margin-left: var(--spacing-xsmall);
-    }
   }
 }
